gcc: linker: use --as-needed if available on Linux
Fedora 30 and other distros are moving to --as-needed on by default.  It only works on gcc + Linux.

Basically it detects listed dependent libs from which no exports are used by the guy including it.

I added the CMake plumbing for it to see what would happen, but at least on Fedora 29 on the configuration I built we are already completely clean for it.

It might show up problems on other platforms or build scenarios I guess... let's see...